Add files via upload
This study analyzed global energy patterns over 20 years and identified countries with maximum and
minimum energy utilization. The results show a positive correlation between the adaptation of renewable
energies of highly populated and larger countries such as United States, India and China and decrease in
their respective CO2 emissions. As a conclusion, we can say that even though renewables are being
utilized all over the world, fossil fuels still occupy a high percentage of the overall energy consumption in
majority of the countries.